+++ changelog.md
@@ -3,6 +3,8 @@
 ## 1.2.0-dev
 
 - Database : ajout de la possibilité de spécifier un ordre par défaut.
+- Foundation/Console : ajout de la commande artisan "optimize:clear".
+- Foundation/Console : suppression de "clear-compiled" de la commande "optimize:all".
 
 ## 1.1.0 (2015-10-26)